Work on ecm_runner.py

+++ client/ecm_runner.py
@@ -3,17 +3,19 @@
 import argparse
 import multiprocessing as mp
 import os
+import random
 import re
+import pprint
 import subprocess
 import time
 
+import dataclasses
 from collections import defaultdict
 from datetime import datetime
-from dataclasses import dataclass
 from typing import List, Tuple
 
 
-@dataclass(frozen=True)
+@dataclasses.dataclass(frozen=True)
 class WorkUnit:
     uid: int
     n: str
@@ -23,29 +25,32 @@ class WorkUnit:
     resume_line: str = ""
 
 
-@dataclass(frozen=True)
+@dataclasses.dataclass(frozen=True)
 class Env:
     ecm_path: str
     extra_params: Tuple[str]
 
 
-@dataclass()
+@dataclasses.dataclass()
 class EcmOutput:
     factors: Tuple[str]
     exit_status: int
     resume_line: str
     using: str
     version: str
-    status: str
+    output: str
+    timings: Tuple[float]
     runtime: float
 
 
 RE_FACTORS_FMT1 = re.compile(r"factor found.*: ([0-9]+)$", re.I | re.MULTILINE)
 RE_FACTORS_FMT2 = re.compile(r"^([0-9]+) [0-9()+*/#!-]", re.I | re.MULTILINE)
-RE_USING = re.compile(r"Using", re.I | re.MULTILINE)
-RE_VERSION = re.compile(r"^GMP-ECM ", re.I | re.MULTILINE)
+RE_USING = re.compile(r"Using .*B1=.*$", re.I | re.MULTILINE)
+RE_VERSION = re.compile(r"^GMP-ECM [0-9].*$", re.I | re.MULTILINE)
 RE_RESUME_N = re.compile(r"\bN=([0-9]+)\b")
 RE_B1_B2 = re.compile(r"\bB1=([0-9]+)\b(.*B2=([0-9]+))?")
+RE_STEP1_TIMING = re.compile(r"Step 1 took ([0-9]+)ms", re.I | re.MULTILINE)
+RE_STEP2_TIMING = re.compile(r"Step 2 took ([0-9]+)ms", re.I | re.MULTILINE)
 
 ECM_ACCEPTED_ARGS = (
     'x0', 'y0', 'param', 'A', 'torsion', 'k', 'power', 'dickson',
@@ -147,12 +152,7 @@ def get_command(wu: WorkUnit, env: Env) -> List[str]:
 
 
 def get_work_units(args, count: int) -> List[WorkUnit]:
-    import random
-
-    if args.resume:
-        units = resume_to_work_units(args, count)
-        args.resume = None
-        return units
+    assert not args.resume
 
     if args.N:
         units = []
@@ -167,7 +167,7 @@ def get_work_units(args, count: int) -> List[WorkUnit]:
     return []
 
 
-def resume_to_work_units(args, count) -> List[WorkUnit]:
+def resume_to_work_units(args) -> List[WorkUnit]:
     last_B1 = None
     with open(args.resume) as f:
         units = []
@@ -177,11 +177,13 @@ def resume_to_work_units(args, count) -> List[WorkUnit]:
 
             match = RE_RESUME_N.search(line)
             assert match, "N not found in resume line: " + repr(line)
-            N = match.group()
+            N = match.group(1)
 
             match = RE_B1_B2.search(line)
             assert match, "B1, B2 not found in resume line"
             B1, _, B2 = match.groups()
+            # TODO B2 should never be found
+            assert B2 is None
             if args.B1 and B1 != args.B1 and B1 != last_B1:
                 last_B1 = B1
                 print()
@@ -207,32 +209,37 @@ def parse_returncode(code: int):
     return code & 1, (code >> 1) & 1, (code >> 2) & 1, (code >> 3) & 1
 
 
-def get_from_stdout(regexp, stdout):
+def _get_match(regexp, stdout):
     match = regexp.search(stdout)
     assert match, (regexp, stdout)
+    if regexp.groups == 1:
+      return match.groups(1)
     return match.group()
 
 
 def process_output(output: subprocess.CompletedProcess):
-    is_error, found_factor, prime_factor, prime_coprime = (
+    is_error, found_factor, prime_factor, prime_cofactor = (
             parse_returncode(output.returncode))
     assert not is_error
 
     factors = tuple()
     if found_factor:
-        print("-" * 80)
-        print(output.stdout)
-        print("-" * 80)
-        print(f"Return: {output.returncode} | {is_error = }, {found_factor = }, {prime_factor = }")
         factors1 = RE_FACTORS_FMT1.findall(output.stdout)
         factors2 = RE_FACTORS_FMT2.findall(output.stdout)
         factors = tuple(sorted(set(map(int, factors1 + factors2))))
-        print("Factor(s):", " ".join(map(str, factors)))
-        print("-" * 80)
-        assert factors
+        # TODO very verbose flag
+        if False:
+            print("-" * 80)
+            print(output.stdout)
+            print("-" * 80)
+            print(f"Return: {output.returncode} | {is_error = }, {found_factor = }, {prime_factor = } {prime_cofactor =}")
+            print("Factor(s):", " ".join(map(str, factors)))
+            print("-" * 80)
 
-    version = get_from_stdout(RE_VERSION, output.stdout)
-    using = get_from_stdout(RE_USING, output.stdout)
+    version = _get_match(RE_VERSION, output.stdout)
+    using = _get_match(RE_USING, output.stdout)
+    step1_timing = _get_match(RE_STEP1_TIMING, output.stdout)
+    step2_timing = _get_match(RE_STEP2_TIMING, output.stdout)
 
     result = EcmOutput(
         factors,
@@ -240,7 +247,8 @@ def process_output(output: subprocess.CompletedProcess):
         resume_line="",
         using=using,
         version=version,
-        status=output.stdout,
+        output=output.stdout,
+        timings=(step1_timing, step2_timing),
         runtime=0)
 
     return result
@@ -274,33 +282,109 @@ def start_workers(env: Env, work: mp.Queue, results: mp.Queue, num_workers: int)
     return workers
 
 
+def print_nth_curve(n):
+    return (
+        (n <= 1) or
+        (n <= 100 and n % 10 == 0) or
+        (n % 100 == 0)
+    )
+
+
+def short_repr(n):
+    n = str(n)
+    if len(n) < 20:
+      return n
+
+    if n.isnumeric():
+      return f"{n[:5]}...{n[-5:]}<{len(n)}>"
+
+    return n
+
+
+def verbose_result_format(wu, result, pp=pprint.PrettyPrinter()):
+    partial = dataclasses.replace(result, output=f"OMMITTED<{len(result.output)}>")
+
+    lines = ["v" * 80]
+    lines.append(pp.pformat(wu).strip())
+    lines.append("-" * 36 + " result " + "-" * 36)
+    lines.append(pp.pformat(partial).strip())
+    lines.append("-" * 36 + " output " + "-" * 36)
+    lines.append(result.output.strip())
+    lines.append("^" * 80)
+    return "\n".join(lines)
+
+
+class ProcessResults:
+  def __init__(self):
+    self.results = defaultdict(list)
+    self.log_fn = f"ecm_runner_{random.randint(0, 99999):5d}.log"
+    self.saved = 0
+    self.f = None
+    self.pp = pprint.PrettyPrinter(width=80, compact=True)
+
+  def _save_result(self, wu, result):
+    if not self.f:
+      print(f"Saving results to {self.log_fn!r}")
+      self.f = open(self.log_fn, "w")
+
+    self.f.write(verbose_result_format(wu, result, pp=self.pp))
+    self.f.write("\n")
+
+
+  def process(self, wu, result):
+    self.results[wu.n].append(result)
+    self.saved += 1
+    count_n = len(self.results[wu.n])
+
+    self._save_result(wu, result)
+
+    if print_nth_curve(count_n):
+        n = short_repr(wu.n)
+        print(f"Result: {self.saved}, Curve: {count_n} N: {n} @ {datetime.now().isoformat()}")
+
+    if result.factors:
+        print(verbose_result_format(wu, result, pp=self.pp))
+        print("Curve count:", count_n)
+        print("Factor(s):", ", ".join(map(str, result.factors)))
+    return result.factors
+
+
 def main_loop(args):
     env = get_env(args)
 
+    # TODO configure with arg?
+    stop_on_factor = True
+
     work = mp.Queue()
     results = mp.Queue()
-    finished = defaultdict(list)
     total_work = 0
     total_finished = 0
+    seen = set()
+
+    process_results = ProcessResults()
 
     workers = start_workers(env, work, results, num_workers=args.threads)
     time.sleep(0.02)
 
+    add_more = True
+
+    if args.resume:
+        units = resume_to_work_units(args)
+        for wu in units:
+            work.put(wu)
+        total_work = len(units)
+        add_more = False
+        print(f"Added {len(units)} work units from -resume {args.resume}")
+
     try:
         while True:
+            time.sleep(0.02)
             while not results.empty():
-                wu, result = results.get_nowait()
                 total_finished += 1
-                finished[wu.n].append(result)
-                count_n = len(finished[wu.n])
+                wu, result = results.get_nowait()
+                process_results.process(wu, result)
 
-                if count_n % 100 == 0:
-                    print("Curves:", count_n, "N:", wu.n, "@", datetime.now().isoformat())
-
-                if result.factors:
-                    print("Result:", result, "from", wu)
-                    print("Curve count:", count_n)
-                    print("FACTOR:", result.factors)
+                if result.factors and stop_on_factor:
                     for worker in workers:
                         worker.terminate()
                     return
@@ -309,36 +393,38 @@ def main_loop(args):
                 assert worker.is_alive()
 
             if work.empty():
-                # TODO get real WorkUnits from server
-                added = 0
-                for wu in get_work_units(args, 2 * args.threads):
-                    added += 1
-                    work.put(wu)
-                    if wu.n not in finished:
-                        # Add to finished
-                        finished[wu.n]
-                        assert wu.n in finished
-                        print("New N:", wu.n)
-                total_work += added
+                if add_more:
+                    # TODO get real WorkUnits from server
+                    added = 0
+                    for wu in get_work_units(args, 2 * args.threads):
+                        added += 1
+                        work.put(wu)
+                        if wu.n not in seen:
+                            seen.add(wu.n)
+                            print("New N:", short_repr(wu.n))
+                    total_work += added
 
-                if added:
-                    print(f"Added {added} work units, finished {total_finished}")
-                else:
-                    print(f"No Work to add: {total_finished}/{total_work}")
-                    if total_work == total_finished:
-                        # No new work, all work finished
-                        assert work.empty()
-                        assert results.empty()
-                        for worker in workers:
-                            if worker:
-                                worker.terminate()
-                        # Done!
-                        return
+                    if added:
+                        print(f"Added {added} work units, finished {total_finished}")
+                    else:
+                        add_more = False
+                        print(f"No work to add: {total_finished}/{total_work}")
 
-            time.sleep(0.02)
+                if total_work == total_finished:
+                    # No new work, all work finished
+                    assert work.empty()
+                    assert results.empty()
+                    for worker in workers:
+                        if worker:
+                            worker.terminate()
+                    # Done!
+                    return
 
     except KeyboardInterrupt:
         print("TODO graceful shutdown in the future")
+        for worker in workers:
+            if worker:
+                worker.terminate()
 
     except:
         print("Main thread had exception!")
